+++ graylog2-server/src/main/java/org/graylog2/system/processing/control/RemoteProcessingControlResource.java
@@ -23,7 +23,7 @@ import retrofit2.http.Path;
 
 import java.util.HashMap;
 
-public interface RemoteProcessingControlResource extends ControlResource {
+public interface RemoteProcessingControlResource {
     @PUT("system/processing/pause")
     Call<Void> pauseProcessing();
